refactor: User 및 History 엔티티 개선 (#15)
- User 엔티티
  - id 타입 Long → Integer 변경
  - @Table(name = "\"user\"") 적용하여 예약어 충돌 회피
- DTO 수정
  - LoginResponse, SignupResponse userId 타입 Long → Integer
- History 엔티티 생성 및 구조 정의

+++ src/main/java/wisoft/backend/auth/dto/LoginResponse.java
@@ -1,13 +1,13 @@
 package wisoft.backend.auth.dto;
 
 public record LoginResponse(
-        Long userId,
+        Integer userId,
         String email,
         String name,
         String message
 ) {
 
-    public static LoginResponse of(Long userId, String email, String name) {
+    public static LoginResponse of(Integer userId, String email, String name) {
         return new LoginResponse(userId, email, name, "로그인에 성공했습니다.");
     }
 }
